August/Essentials_LLM/NouSmartLibrarian/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-dotenv==1.0.1
redis==5.0.4
email-validator==2.1.1
aiofiles==23.2.1
openai==1.30.1
chromadb==1.0.20
httpx==0.27.0
regex==2024.11.6
loguru==0.7.3
tiktoken==0.7.0